+++ internal/web/errors.go
@@ -1,0 +1,5 @@
+package web
+
+import "errors"
+
+var ErrWebDisabled = errors.New("web server is disabled")